openmp linking with cmake and github actions...no source files changed

+++ compile.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
+if [[ $# -eq 0 ]] ; then
+	echo 'Run with one argument: 1 or 0. If 1, OpenMP is enabled. If 0, OpenMP is disabled.'
+	echo 'USAGE: ./compile.sh 1'
+	exit 0
+fi
 
 echo "Cleaning old files..."
 rm -f ljmd*.x
@@ -6,10 +11,14 @@ rm -rf build/
 
 echo "Preparing to compile with CMake..."
 mkdir -p build
-cmake -S. -B ./build -D ENABLE_TESTING=yes -DCMAKE_EXPORT_COMPILE_COMMANDS=1
+cmake -S. -B ./build -D ENABLE_TESTING=yes -DCMAKE_EXPORT_COMPILE_COMMANDS=1 -DENABLE_OPENMP=$1
 cmake --build build/
 
+export OMP_DYNAMIC=FALSE
+export OMP_NUM_THREADS=4
 ./test_verification.sh
 
-# ./result_verification.sh
+export OMP_DYNAMIC=FALSE
+export OMP_NUM_THREADS=4
+./result_verification.sh
